Prerelease v1.0.0a2

+++ src/localis/data/database.py
@@ -22,7 +22,6 @@ class Database:
         self._conn = sqlite3.connect(self.db_path)
         self._conn.row_factory = sqlite3.Row
 
-        self._conn.execute("PRAGMA journal_mode = OFF")
         self._conn.execute("PRAGMA synchronous = OFF")
         self._conn.execute("PRAGMA temp_store = MEMORY")
         self._conn.execute("PRAGMA cache_size = -100000")
@@ -47,9 +46,6 @@ class Database:
     def commit(self) -> None:
         self._conn.commit()
 
-    def clone(self) -> "Database":
-        return Database(self.db_path)
-
     @contextmanager
     def atomic(self):
         try:
@@ -65,8 +61,8 @@ class Database:
         path = Path(dir) if dir else Path.cwd()
         target_path = path / filename
 
-        with resources.path(cls.PATH, cls.FILENAME) as db_file:
-            shutil.copy(db_file, target_path)
+        db_file = resources.files(cls.PATH) / cls.FILENAME
+        shutil.copy(db_file, target_path)
 
         return str(target_path)
 
@@ -112,6 +108,9 @@ class Database:
         for table in tables:
             if hasattr(table, "drop") and callable(table.drop):
                 table.drop()
+            elif isinstance(table, str):
+                db.execute(f"DROP TABLE {table}")
+        self.commit()
 
     def create_fts_table(
         self,
@@ -138,13 +137,10 @@ class Database:
         """Vacuum database"""
         self.execute("VACUUM")
 
-    def analyze(self) -> sqlite3.Cursor:
-        """Analyze database for query optimization"""
-        return self.execute("ANALYZE")
-
     def set_db_path(self, path: str) -> None:
         """Stores the database path in a config file for external storage and reloads the connection with the new path."""
-        self.CONFIG_FILE.write_text(path)
+        if path != ":memory:":
+            self.CONFIG_FILE.write_text(path)
         self.close()
         self.db_path = path
         self._setup_conn()
@@ -155,18 +151,24 @@ class Database:
         if cls.CONFIG_FILE.exists():
             return cls.CONFIG_FILE.read_text().strip()
 
-        from importlib.resources import files
-
-        db_file = files(cls.PATH) / cls.FILENAME
+        db_file = resources.files(cls.PATH) / cls.FILENAME
         return str(db_file)
 
     def revert_to_default(self):
         """Removes config file, external database and reverts to the bundled db file."""
         self.close()
-        self.CONFIG_FILE.unlink()
+
+        if self.CONFIG_FILE.exists():
+            self.CONFIG_FILE.unlink()
+
         db_file = Path(self.db_path)
-        db_file.unlink()
-        self.db_path = self.get_db_path()
+
+        new_path = self.get_db_path()
+
+        if db_file.exists() and str(db_file) != new_path:
+            db_file.unlink()
+
+        self.db_path = new_path
         self._setup_conn()
